更改鉴权方式为注解鉴权

+++ sp-server/src/main/java/com/pj/project4sp/admin/SpAdmin.java
@@ -36,7 +36,7 @@ public class SpAdmin implements Serializable  {
 	public String phone;	
 	
 	/** 所属角色id */
-	public Integer roleId;	
+	public Long roleId;	
 	
 	/** 账号状态(1=正常, 2=禁用) */
 	public Integer status;
